changes to flags single sorce
